Moved these files into an account folder because they are all related to making/viewing/editing an account.
